+++ src/test/kotlin/org/jetbrains/changelog/tasks/PatchChangelogTaskTest.kt
@@ -2,6 +2,8 @@ package org.jetbrains.changelog.tasks
 
 import org.jetbrains.changelog.BaseTest
 import org.jetbrains.changelog.exceptions.MissingVersionException
+import java.text.SimpleDateFormat
+import java.util.Date
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -89,6 +91,46 @@ class PatchChangelogTaskTest : BaseTest() {
         runTask("patchChangelog")
 
         assertEquals("## Foo 1.0.0 bar buz", extension.get().getHeader())
+    }
+
+    @Test
+    fun `applies custom header with date`() {
+        changelog = """
+            # Changelog
+            All notable changes to this project will be documented in this file.
+
+            ## [Unreleased]
+
+            ### Added
+            - Some other thing added.
+
+            ## [1.0.0] - 2020-07-02
+
+            ### Added
+            - Something added.
+        """.trimIndent()
+        buildFile = """
+            import java.text.SimpleDateFormat
+            import java.util.Arrays
+            import java.util.Date
+
+            plugins {
+                id 'org.jetbrains.changelog'
+            }
+            changelog {
+                version = "1.0.0"
+                
+                headerFormat = "[{0}] - {1}"
+                headerArguments = Arrays.asList(version, new SimpleDateFormat("yyyy-MM-dd").format(new Date()))
+            }
+        """.trimIndent()
+        extension.headerFormat = "[{0}] - {1}"
+
+        project.evaluate()
+        runTask("patchChangelog")
+
+        val date = SimpleDateFormat("yyyy-MM-dd").format(Date())
+        assertEquals("## [1.0.0] - $date", extension.get().getHeader())
     }
 
     @Test